+++ src/index.js
@@ -1,6 +1,6 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-import "./globalStyles.css";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import './global.css';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById('root'));
